requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Embeddings (local MLX — Apple Silicon optimized)
mlx-embeddings>=0.0.5

# Database
sqlalchemy>=2.0.0
pgvector>=0.2.0
psycopg2-binary>=2.9.9

# MCP
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
